+++ src/signature.ts
@@ -39,12 +39,6 @@ const stateEffect = StateEffect.define<SignatureEffect>(),
 		},
 	});
 
-/**
- * Get the [signatureHelp](https://github.com/bhsd-harry/codemirror-mediawiki/tree/wikitext#signaturehelp)
- * extension for Wikitext.
- * @param configData [WikiParser-Node](https://www.npmjs.com/package/wikiparser-node) configuration data.
- * @param cdn [jsDelivr CDN](https://www.jsdelivr.com/network), defaulting to `https://testingcf.jsdelivr.net`
- */
 export default (
 	configData: ConfigData,
 	cdn?: string,